move grouping to outputtransform
